refactor: 경로 별칭 반영
