+++ src/shared/components/index.ts
@@ -1,0 +1,4 @@
+export { default as CommonIcon } from './Icon';
+export { default as CommonText } from './Text';
+export { default as CommonAvatar } from './Avatar';
+export { default as CommonBadge } from './Badge';
